Instructions & Credits created
Feature: Credits Screen created

Corrections for Pylint

Pylint Corrections

Quitting button fixed

Pylint and comment corrections done

Pylint Corrections

Credits Screen created

add original text and fix quit button

Entity and Enemy

squash

Movement fixed

Movement fixed again

fix float to int suspicious conversion

implementation dijkstra

implementation dijkstra OK

level diagonal ok, so graph with floats

fix clarity and optimisation

textures of the world

fak

fix toml

+++ src/core/dijkstra.py
@@ -26,7 +26,7 @@ class Graph:
     """
     Graph class to represent a graph, with nodes and edges
     """
-    def __init__(self, edges: dict[str, dict[str, int]]):
+    def __init__(self, edges: dict[str, dict[str, float]]):
         self.edges = edges
         self.nodes = {node: Node(node) for node in edges.keys()}
